feat #28: label, milestone, joinUser를 불러오는 API 작성
처음에만 모든 정보(issues, labels, milestones, joinUsers)를 불러오기 때문에
각각 따로 가져오는 API가 아닌, 한번에 이 모두를 불러오는 API를 작성하였다.
따라서 이전에 있던 Issues를 불러오는 API는 삭제했다.
