+++ backend/app.js
@@ -8,7 +8,7 @@ require('dotenv').config();
 const { sequelize } = require('./models');
 const indexRouter = require('./routes/index');
 const authRouter = require('./routes/auth');
-const issuesRouter = require('./routes/issues');
+const allDataRouter = require('./routes/allData');
 const issueRouter = require('./routes/issue');
 const session = require('express-session');
 const cors = require('cors');
@@ -51,7 +51,7 @@ passportConfig();
 app.use('/api', indexRouter);
 app.use('/api/auth', authRouter);
 app.use('/api/issue', issueRouter);
-app.use('/api/issues', issuesRouter);
+app.use('/api/allData', allDataRouter);
 
 process.env.ENV === 'production' &&
   app.get('*', (req, res) => {
